Stashing changes

+++ CreatureEngine/Profiling/RenderUtilities.h
@@ -90,7 +90,7 @@ namespace Profiling
                 }
             }
 
-            ColorTemplate = new Pixel[Size.x];
+            ColorTemplate = new Pixel[(size_t)Size.x];
             for_loop(x, Size.x)
             {
                 ColorTemplate[x] = Pixel(255, 0, 0, 150);
@@ -183,7 +183,7 @@ namespace Profiling
 			    PreviousX = val;
 
                 float Y = (_value * (_value / DataRange.x)) * (1.0f / Size.x);
-                val = std::log(_value) * Size.x / std::log(DataRange.x);
+                val = static_cast<int>(std::log(_value) * Size.x / std::log(DataRange.x));
 				swapBuffer();
                 DisplayTexture.Update((uint8_t*)ReadBuffer);
             }
